Add an envelope type for workflow runs

+++ pkg/manager/api/workflowrun_test.go
@@ -71,6 +71,6 @@ func TestWorkflowRunManager(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Equal(t, workflowName, wr.Name)
-	require.Equal(t, int32(2), wr.RunNumber)
+	require.Equal(t, 2, wr.RunNumber)
 	require.Equal(t, fmt.Sprintf("/workflows/%s/runs/2", workflowName), wr.AppURL.Path)
 }
